+++ packages/verified-fetch/test/utils/parse-url-string.spec.ts
@@ -3,10 +3,13 @@ import { defaultLogger } from '@libp2p/logger'
 import { createEd25519PeerId } from '@libp2p/peer-id-factory'
 import { expect } from 'aegir/chai'
 import { CID } from 'multiformats/cid'
+import { match } from 'sinon'
 import { stubInterface } from 'sinon-ts'
 import { parseUrlString } from '../../src/utils/parse-url-string.js'
 import type { IPNS } from '@helia/ipns'
 import type { ComponentLogger, PeerId } from '@libp2p/interface'
+import type { Answer } from '@multiformats/dns'
+import type { IPNSRecordV2 } from 'ipns'
 import type { StubbedInstance } from 'sinon-ts'
 
 const HTTP_PROTOCOLS = [
@@ -154,7 +157,7 @@ describe('parseUrlString', () => {
   describe('ipns://<dnsLinkDomain> URLs', () => {
     it('handles invalid DNSLinkDomains', async () => {
       ipns.resolve.rejects(new Error('Unexpected failure from ipns resolve method'))
-      ipns.resolveDns.rejects(new Error('Unexpected failure from ipns dns query'))
+      ipns.resolveDNSLink.rejects(new Error('Unexpected failure from ipns dns query'))
 
       await expect(parseUrlString({ urlString: 'ipns://mydomain.com', ipns, logger })).to.eventually.be.rejected
         .with.property('errors').that.deep.equals([
@@ -164,9 +167,10 @@ describe('parseUrlString', () => {
     })
 
     it('can parse a URL with DNSLinkDomain only', async () => {
-      ipns.resolveDns.withArgs('mydomain.com').resolves({
+      ipns.resolveDNSLink.withArgs('mydomain.com').resolves({
         cid: CID.parse('QmQJ8fxavY54CUsxMSx9aE9Rdcmvhx8awJK2jzJp4iAqCr'),
-        path: ''
+        path: '',
+        answer: stubInterface<Answer>()
       })
 
       await assertMatchUrl(
@@ -180,9 +184,10 @@ describe('parseUrlString', () => {
     })
 
     it('can parse a URL with DNSLinkDomain+path', async () => {
-      ipns.resolveDns.withArgs('mydomain.com').resolves({
+      ipns.resolveDNSLink.withArgs('mydomain.com').resolves({
         cid: CID.parse('QmQJ8fxavY54CUsxMSx9aE9Rdcmvhx8awJK2jzJp4iAqCr'),
-        path: ''
+        path: '',
+        answer: stubInterface<Answer>()
       })
 
       await assertMatchUrl(
@@ -196,9 +201,10 @@ describe('parseUrlString', () => {
     })
 
     it('can parse a URL with DNSLinkDomain+queryString', async () => {
-      ipns.resolveDns.withArgs('mydomain.com').resolves({
+      ipns.resolveDNSLink.withArgs('mydomain.com').resolves({
         cid: CID.parse('QmQJ8fxavY54CUsxMSx9aE9Rdcmvhx8awJK2jzJp4iAqCr'),
-        path: ''
+        path: '',
+        answer: stubInterface<Answer>()
       })
 
       await assertMatchUrl(
@@ -214,9 +220,10 @@ describe('parseUrlString', () => {
     })
 
     it('can parse a URL with DNSLinkDomain+path+queryString', async () => {
-      ipns.resolveDns.withArgs('mydomain.com').resolves({
+      ipns.resolveDNSLink.withArgs('mydomain.com').resolves({
         cid: CID.parse('QmQJ8fxavY54CUsxMSx9aE9Rdcmvhx8awJK2jzJp4iAqCr'),
-        path: ''
+        path: '',
+        answer: stubInterface<Answer>()
       })
 
       await assertMatchUrl(
@@ -232,9 +239,10 @@ describe('parseUrlString', () => {
     })
 
     it('can parse a URL with DNSLinkDomain+directoryPath+queryString', async () => {
-      ipns.resolveDns.withArgs('mydomain.com').resolves({
+      ipns.resolveDNSLink.withArgs('mydomain.com').resolves({
         cid: CID.parse('QmQJ8fxavY54CUsxMSx9aE9Rdcmvhx8awJK2jzJp4iAqCr'),
-        path: ''
+        path: '',
+        answer: stubInterface<Answer>()
       })
 
       await assertMatchUrl(
@@ -409,7 +417,7 @@ describe('parseUrlString', () => {
 
     it('handles invalid PeerIds', async () => {
       ipns.resolve.rejects(new Error('Unexpected failure from ipns resolve method'))
-      ipns.resolveDns.rejects(new Error('Unexpected failure from ipns dns query'))
+      ipns.resolveDNSLink.rejects(new Error('Unexpected failure from ipns dns query'))
 
       await expect(parseUrlString({ urlString: 'ipns://123PeerIdIsFake456', ipns, logger })).to.eventually.be.rejected
         .with.property('errors').that.deep.equals([
@@ -420,7 +428,7 @@ describe('parseUrlString', () => {
 
     it('handles valid PeerId resolve failures', async () => {
       ipns.resolve.rejects(new Error('Unexpected failure from ipns resolve method'))
-      ipns.resolveDns.rejects(new Error('Unexpected failure from ipns dns query'))
+      ipns.resolveDNSLink.rejects(new Error('Unexpected failure from ipns dns query'))
 
       await expect(parseUrlString({ urlString: `ipns://${testPeerId}`, ipns, logger })).to.eventually.be.rejected
         .with.property('errors').that.deep.equals([
@@ -432,7 +440,8 @@ describe('parseUrlString', () => {
     it('can parse a URL with PeerId only', async () => {
       ipns.resolve.withArgs(matchPeerId(testPeerId)).resolves({
         cid: CID.parse('QmQJ8fxavY54CUsxMSx9aE9Rdcmvhx8awJK2jzJp4iAqCr'),
-        path: ''
+        path: '',
+        record: stubInterface<IPNSRecordV2>()
       })
 
       await assertMatchUrl(
@@ -448,7 +457,8 @@ describe('parseUrlString', () => {
     it('can parse a URL with PeerId+path', async () => {
       ipns.resolve.withArgs(matchPeerId(testPeerId)).resolves({
         cid: CID.parse('QmQJ8fxavY54CUsxMSx9aE9Rdcmvhx8awJK2jzJp4iAqCr'),
-        path: ''
+        path: '',
+        record: stubInterface<IPNSRecordV2>()
       })
 
       await assertMatchUrl(
@@ -464,7 +474,8 @@ describe('parseUrlString', () => {
     it('can parse a URL with PeerId+path with a trailing slash', async () => {
       ipns.resolve.withArgs(matchPeerId(testPeerId)).resolves({
         cid: CID.parse('QmQJ8fxavY54CUsxMSx9aE9Rdcmvhx8awJK2jzJp4iAqCr'),
-        path: ''
+        path: '',
+        record: stubInterface<IPNSRecordV2>()
       })
 
       await assertMatchUrl(
@@ -480,7 +491,8 @@ describe('parseUrlString', () => {
     it('can parse a URL with PeerId+queryString', async () => {
       ipns.resolve.withArgs(matchPeerId(testPeerId)).resolves({
         cid: CID.parse('QmQJ8fxavY54CUsxMSx9aE9Rdcmvhx8awJK2jzJp4iAqCr'),
-        path: ''
+        path: '',
+        record: stubInterface<IPNSRecordV2>()
       })
 
       await assertMatchUrl(
@@ -498,7 +510,8 @@ describe('parseUrlString', () => {
     it('can parse a URL with PeerId+path+queryString', async () => {
       ipns.resolve.withArgs(matchPeerId(testPeerId)).resolves({
         cid: CID.parse('QmQJ8fxavY54CUsxMSx9aE9Rdcmvhx8awJK2jzJp4iAqCr'),
-        path: ''
+        path: '',
+        record: stubInterface<IPNSRecordV2>()
       })
 
       await assertMatchUrl(
@@ -521,7 +534,8 @@ describe('parseUrlString', () => {
 
       ipns.resolve.withArgs(matchPeerId(peerId)).resolves({
         cid,
-        path: recordPath
+        path: recordPath,
+        record: stubInterface<IPNSRecordV2>()
       })
 
       await assertMatchUrl(
@@ -542,7 +556,8 @@ describe('parseUrlString', () => {
 
       ipns.resolve.withArgs(matchPeerId(peerId)).resolves({
         cid,
-        path: recordPath
+        path: recordPath,
+        record: stubInterface<IPNSRecordV2>()
       })
 
       await assertMatchUrl(
@@ -563,7 +578,8 @@ describe('parseUrlString', () => {
 
       ipns.resolve.withArgs(matchPeerId(peerId)).resolves({
         cid,
-        path: recordPath
+        path: recordPath,
+        record: stubInterface<IPNSRecordV2>()
       })
 
       await assertMatchUrl(
@@ -586,7 +602,8 @@ describe('parseUrlString', () => {
       cid = CID.parse('QmdmQXB2mzChmMeKY47C43LxUdg1NDJ5MWcKMKxDu7RgQm')
       ipns.resolve.withArgs(matchPeerId(peerId)).resolves({
         cid,
-        path: ''
+        path: '',
+        record: stubInterface<IPNSRecordV2>()
       })
     })
 
@@ -673,7 +690,8 @@ describe('parseUrlString', () => {
         cid = CID.parse('QmdmQXB2mzChmMeKY47C43LxUdg1NDJ5MWcKMKxDu7RgQm')
         ipns.resolve.withArgs(matchPeerId(peerId)).resolves({
           cid,
-          path: ''
+          path: '',
+          record: stubInterface<IPNSRecordV2>()
         })
       })
 
@@ -751,23 +769,51 @@ describe('parseUrlString', () => {
     })
   })
 
-  HTTP_PROTOCOLS.forEach(proto => {
-    describe(`${proto}://<key>.ipns.example.com URLs`, () => {
-      let peerId: PeerId
-      let cid: CID
+  const IPNS_TYPES = [
+    ['dnslink-encoded', (i: number) => `${i}-example-com`],
+    ['dnslink-decoded', (i: number) => `${i}.example.com`],
+    ['peerid', async () => createEd25519PeerId()]
+  ] as const
 
+  IPNS_TYPES.flatMap(([type, fn]) => {
+    // merge IPNS_TYPES with HTTP_PROTOCOLS
+    return HTTP_PROTOCOLS.reduce<Array<[string, string, (i: number) => string | Promise<PeerId>]>>((acc, proto) => {
+      acc.push([proto, type, fn])
+      return acc
+    }, [])
+  }, []).forEach(([proto, type, getVal]) => {
+    describe(`${proto}://<${type}>.ipns.example.com URLs`, () => {
+      let value: PeerId | string
+      let cid: CID
+      let i = 0
       beforeEach(async () => {
-        peerId = await createEd25519PeerId()
+        value = await getVal(i++)
         cid = CID.parse('QmdmQXB2mzChmMeKY47C43LxUdg1NDJ5MWcKMKxDu7RgQm')
-        ipns.resolve.withArgs(matchPeerId(peerId)).resolves({
-          cid,
-          path: ''
-        })
+        if (type === 'peerid') {
+          ipns.resolve.withArgs(matchPeerId(value as PeerId)).resolves({
+            cid,
+            path: '',
+            record: stubInterface<IPNSRecordV2>()
+          })
+        } else if (type === 'dnslink-encoded') {
+          const matchValue = (value as string).replace(/-/g, '.')
+          ipns.resolveDNSLink.withArgs(match(matchValue)).resolves({
+            cid,
+            path: '',
+            answer: stubInterface<Answer>()
+          })
+        } else {
+          ipns.resolveDNSLink.withArgs(match(value as string)).resolves({
+            cid,
+            path: '',
+            answer: stubInterface<Answer>()
+          })
+        }
       })
 
       it('should parse a IPNS Subdomain Gateway URL with a CID only', async () => {
         await assertMatchUrl(
-          `${proto}://${peerId}.ipns.example.com`, {
+          `${proto}://${value.toString()}.ipns.example.com`, {
             protocol: 'ipns',
             cid: cid.toString(),
             path: '',
@@ -778,7 +824,7 @@ describe('parseUrlString', () => {
 
       it('can parse a IPNS Subdomain Gateway URL with CID+path', async () => {
         await assertMatchUrl(
-          `${proto}://${peerId}.ipns.example.com/1 - Barrel - Part 1/1 - Barrel - Part 1 - alt.txt`, {
+          `${proto}://${value.toString()}.ipns.example.com/1 - Barrel - Part 1/1 - Barrel - Part 1 - alt.txt`, {
             protocol: 'ipns',
             cid: cid.toString(),
             path: '1 - Barrel - Part 1/1 - Barrel - Part 1 - alt.txt',
@@ -789,7 +835,7 @@ describe('parseUrlString', () => {
 
       it('can parse a IPNS Subdomain Gateway URL with CID+directoryPath', async () => {
         await assertMatchUrl(
-          `${proto}://${peerId}.ipns.example.com/path/to/dir/`, {
+          `${proto}://${value.toString()}.ipns.example.com/path/to/dir/`, {
             protocol: 'ipns',
             cid: cid.toString(),
             path: 'path/to/dir/',
@@ -800,7 +846,7 @@ describe('parseUrlString', () => {
 
       it('can parse a IPNS Subdomain Gateway URL with CID+queryString', async () => {
         await assertMatchUrl(
-          `${proto}://${peerId}.ipns.example.com?format=car`, {
+          `${proto}://${value.toString()}.ipns.example.com?format=car`, {
             protocol: 'ipns',
             cid: cid.toString(),
             path: '',
@@ -813,7 +859,7 @@ describe('parseUrlString', () => {
 
       it('can parse a IPNS Subdomain Gateway URL with CID+path+queryString', async () => {
         await assertMatchUrl(
-          `${proto}://${peerId}.ipns.example.com/1 - Barrel - Part 1/1 - Barrel - Part 1 - alt.txt?format=tar`, {
+          `${proto}://${value.toString()}.ipns.example.com/1 - Barrel - Part 1/1 - Barrel - Part 1 - alt.txt?format=tar`, {
             protocol: 'ipns',
             cid: cid.toString(),
             path: '1 - Barrel - Part 1/1 - Barrel - Part 1 - alt.txt',
@@ -826,7 +872,7 @@ describe('parseUrlString', () => {
 
       it('can parse a IPNS Subdomain Gateway URL with CID+directoryPath+queryString', async () => {
         await assertMatchUrl(
-          `${proto}://${peerId}.ipns.example.com/path/to/dir/?format=tar`, {
+          `${proto}://${value.toString()}.ipns.example.com/path/to/dir/?format=tar`, {
             protocol: 'ipns',
             cid: cid.toString(),
             path: 'path/to/dir/',